adding initial version of controller/collector utilities... also merging patch from Jesse Norell <jesse@kci.net> which improved the error checking...and added support for udp connected sockets.  Not sure this is the way to go but including it in svn for now.
The primary user visible impact will be that the 
collector will have to be started first before
the tap.

Also be warned that right now the "process registry" has an 
important bug pending that does not check to see if the 
exec of the "tap" worked or not.  This can result in 
client/server getting out of sync.

The early commit is mainly to get some user input on the 
design/interface...