bug fix to make the -f "filter" command work correctly...
the collector was not able to parse the -f option correctly as sent
from the controller....this now works the right format is:

"-x x -y y -z z -f " host 1.1.1.1 and port 80 " "

note the spaces.  I will fix the requirement of the extra spaces in 
the next commit.  

Also, a bug fix to check to make sure when we get a request to kill 
a process that we actually started that process and we are not killing
any arbitrary process.