merging socket patch from Jesse Norell <jesse@kci.net> "   This one is against rev14.  It makes "-i interface" to tap.c optional (and removes the hard-coded -i addition to the tap command invocation in collector) and makes the "-d dest" accept either a hostname or ip address (and if the hostname has multiple ip addrs, it'll loop through them all trying to connect to one), and allows port numbers to be names or numbers.
  TCP support probably isn't far away .. if you just change "SOCK_DGRAM"
to "SOCK_STREAM" it'll try to use tcp, but I don't know that it'll work
right (I think we either need to use transactional tcp, which I know
nothing about, or be able to mark the start/stop of individual "packets"
within the tcp stream).

  I'll probably look at adding logging/debugging to some of the other
programs next.  Are you working on any of them right now, Manish?  If
so, I'll avoid the ones you're working on for now to avoid messy patch
integration later.
"